@@ -24,13 +24,13 @@ class Animal {
   final String name;
 
   Animal({
-    this.id,
-    this.name,
+    required this.id,
+    required this.name,
   });
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
   final String title;
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -67,13 +67,13 @@ class _MyHomePageState extends State<MyHomePage> {
     Animal(id: 27, name: "Dolphin"),
   ];
   final _items = _animals
-      .map((animal) => MultiSelectItem<Animal>(animal, animal.name))
+      .map((animal) => MultiSelectItem<Animal?>(animal, animal.name))
       .toList();
   //List<Animal> _selectedAnimals = [];
-  List<Animal> _selectedAnimals2 = [];
-  List<Animal> _selectedAnimals3 = [];
+  List<Animal?> _selectedAnimals2 = [];
+  List<Animal?> _selectedAnimals3 = [];
   //List<Animal> _selectedAnimals4 = [];
-  List<Animal> _selectedAnimals5 = [];
+  List<Animal?> _selectedAnimals5 = [];
   final _multiSelectKey = GlobalKey<FormFieldState>();
 
   @override
@@ -148,7 +148,7 @@ class _MyHomePageState extends State<MyHomePage> {
                       title: Text("Animals"),
                       items: _items,
                       onConfirm: (values) {
-                        _selectedAnimals2 = values;
+                        _selectedAnimals2 = values as List<Animal?>;
                       },
                       chipDisplay: MultiSelectChipDisplay(
                         onTap: (value) {
@@ -158,7 +158,7 @@ class _MyHomePageState extends State<MyHomePage> {
                         },
                       ),
                     ),
-                    _selectedAnimals2 == null || _selectedAnimals2.isEmpty
+                    _selectedAnimals2.isEmpty
                         ? Container(
                             padding: EdgeInsets.all(10),
                             alignment: Alignment.centerLeft,
@@ -174,7 +174,7 @@ class _MyHomePageState extends State<MyHomePage> {
               //################################################################################################
               // MultiSelectBottomSheetField with validators
               //################################################################################################
-              MultiSelectBottomSheetField<Animal>(
+              MultiSelectBottomSheetField<Animal?>(
                 key: _multiSelectKey,
                 initialChildSize: 0.7,
                 maxChildSize: 0.95,
@@ -186,7 +186,7 @@ class _MyHomePageState extends State<MyHomePage> {
                   if (values == null || values.isEmpty) {
                     return "Required";
                   }
-                  List<String> names = values.map((e) => e.name).toList();
+                  List<String> names = values.map((e) => e!.name).toList();
                   if (names.contains("Frog")) {
                     return "Frogs are weird!";
                   }
@@ -196,14 +196,14 @@ class _MyHomePageState extends State<MyHomePage> {
                   setState(() {
                     _selectedAnimals3 = values;
                   });
-                  _multiSelectKey.currentState.validate();
+                  _multiSelectKey.currentState?.validate();
                 },
                 chipDisplay: MultiSelectChipDisplay(
                   onTap: (item) {
                     setState(() {
                       _selectedAnimals3.remove(item);
                     });
-                    _multiSelectKey.currentState.validate();
+                    _multiSelectKey.currentState?.validate();
                   },
                 ),
               ),
@@ -217,7 +217,7 @@ class _MyHomePageState extends State<MyHomePage> {
                 title: Text("Animals"),
                 headerColor: Colors.blue.withOpacity(0.5),
                 decoration: BoxDecoration(
-                  border: Border.all(color: Colors.blue[700], width: 1.8),
+                  border: Border.all(color: Colors.blue[700]!, width: 1.8),
                 ),
                 selectedChipColor: Colors.blue.withOpacity(0.5),
                 selectedTextStyle: TextStyle(color: Colors.blue[800]),
@@ -231,7 +231,7 @@ class _MyHomePageState extends State<MyHomePage> {
               //################################################################################################
               MultiSelectDialogField(
                 onConfirm: (val) {
-                  _selectedAnimals5 = val;
+                  _selectedAnimals5 = val as List<Animal>;
                 },
                 items: _items,
                 initialValue:
